Move util/dev/validate to util/valid-config
